All cash is in longs. Fixed crash in a few Journal entries. Cleaned up pretty print and added lat/long conversion.
Added an index of name only for materialscommodities for future..
Material commodities now uses long.  Journal does not use AutoSizeRows.

+++ EDDiscovery/EliteDangerous/JournalEvents/JournalBuyAmmo.cs
@@ -11,10 +11,10 @@ namespace EDDiscovery.EliteDangerous.JournalEvents
     {
         public JournalBuyAmmo(JObject evt ) : base(evt, JournalTypeEnum.BuyAmmo)
         {
-            Cost = Tools.GetInt(evt["Cost"]);
+            Cost = Tools.GetLong(evt["Cost"]);
 
         }
-        public int Cost { get; set; }
+        public long Cost { get; set; }
 
         public static System.Drawing.Bitmap Icon { get { return EDDiscovery.Properties.Resources.ammunition; } }
